removing transaction from link when it isnt defined
