+++ x-pack/plugins/apm/public/components/shared/charts/Timeline/Marker/index.test.tsx
@@ -6,9 +6,9 @@
 
 import { shallow } from 'enzyme';
 import React from 'react';
-import { Marker } from '../';
-import { AgentMark } from '../../../../../app/TransactionDetails/WaterfallWithSummmary/WaterfallContainer/Marks/get_agent_marks';
-import { ErrorMark } from '../../../../../app/TransactionDetails/WaterfallWithSummmary/WaterfallContainer/Marks/get_error_marks';
+import { Marker } from './';
+import { AgentMark } from '../../../../app/TransactionDetails/WaterfallWithSummmary/WaterfallContainer/Marks/get_agent_marks';
+import { ErrorMark } from '../../../../app/TransactionDetails/WaterfallWithSummmary/WaterfallContainer/Marks/get_error_marks';
 
 describe('Marker', () => {
   it('renders agent marker', () => {
